requirements/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.10.0=py37hc8dfbb8_1
aiohttp=3.6.2=py37h8f50634_1
altair=4.1.0=pypi_0
argh=0.26.2=pyh9f0ad1d_1002
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=pypi_0
async-generator=1.10=pypi_0
async-timeout=3.0.1=py_1000
attrs=20.2.0=pyh9f0ad1d_0
backcall=0.2.0=py_0
base58=2.0.1=pypi_0
blas=1.0=mkl
bleach=3.2.1=pypi_0
blinker=1.4=pypi_0
boto=2.49.0=py37_0
boto3=1.15.16=pypi_0
botocore=1.18.16=pypi_0
brotlipy=0.7.0=py37hb5d75c8_1001
c-ares=1.16.1=h516909a_3
ca-certificates=2021.1.19=h06a4308_0
cachetools=4.1.1=py_0
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.3=py37he30daa8_0
chardet=3.0.4=py37he5f6b98_1008
click=7.1.2=pyh9f0ad1d_0
configparser=5.0.1=py_0
cryptography=3.1.1=py37hb09aad4_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py_2
decorator=4.4.2=py_0
defusedxml=0.6.0=pypi_0
dgl-cuda10.2=0.5.2=py37_0
dill=0.3.2=pyh9f0ad1d_0
docker-pycreds=0.4.0=py_0
emoji=0.6.0=pyh9f0ad1d_0
entrypoints=0.3=pypi_0
enum-compat=0.0.3=pypi_0
filelock=3.0.12=pypi_0
freetype=2.10.3=h5ab3b9f_0
fsspec=0.8.3=py_0
future=0.18.2=py37hc8dfbb8_2
gensim=3.8.0=py37h962f231_0
gitdb=4.0.5=py_0
gitpython=3.1.9=py_0
google-api-core=1.22.2=py37_0
google-auth=1.22.0=py_0
google-auth-oauthlib=0.4.1=py_2
google-cloud-core=1.6.0=pyhd3eb1b0_0
google-cloud-storage=1.31.0=py_0
google-crc32c=1.1.2=py37h27cfd23_0
google-resumable-media=1.2.0=pyhd3eb1b0_1
googleapis-common-protos=1.52.0=py37_0
gql=0.1.0=py_0
graphql-core=3.1.2=py_0
grpcio=1.31.0=py37hb0870dc_0
idna=2.10=pyh9f0ad1d_0
imbalanced-learn=0.7.0=py_1
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
ipykernel=5.3.4=pypi_0
ipython=7.18.1=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.2=py37_0
jinja2=2.11.2=pypi_0
jmespath=0.10.0=py_0
joblib=0.17.0=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
kiwisolver=1.2.0=py37h99015e2_1
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcrc32c=1.1.1=he6710b0_2
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=hae1eefd_17
libgfortran4=7.5.0=hae1eefd_17
libllvm10=10.0.1=he513fc3_3
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=h8b12597_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
llvmlite=0.34.0=py37h269e1b5_4
lz4-c=1.9.2=heb0550a_3
markdown=3.3.1=pyh9f0ad1d_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37hd478181_0
mistune=0.8.4=pypi_0
mittens=0.2=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
multidict=4.7.5=py37h8f50634_2
nbclient=0.5.0=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.0.7=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=pypi_0
networkx=2.5=py_0
ninja=1.10.1=py37hfd86e86_0
nltk=3.5=py_0
notebook=6.1.4=pypi_0
numba=0.51.2=py37h9fdb41a_0
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
nvidia-ml=7.352.0=py_0
oauthlib=3.0.1=py_0
olefile=0.46=py37_0
openssl=1.1.1i=h27cfd23_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.3=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.7.0=py_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=py37_1
pickleshare=0.7.5=py37_1001
pillow=7.2.0=py37hb39fc2d_0
pip=20.2.3=py37_0
prometheus-client=0.8.0=pypi_0
promise=2.3=py37hc8dfbb8_2
prompt-toolkit=3.0.7=py_0
protobuf=3.13.0=pypi_0
psutil=5.7.2=py37hb5d75c8_1
ptyprocess=0.6.0=py37_0
pyarrow=1.0.1=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pydeck=0.5.0b1=pypi_0
pygments=2.7.1=py_0
pyjwt=1.7.1=py_0
pynndescent=0.5.1=pyhd3deb0d_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=pypi_0
pysocks=1.7.1=py37he5f6b98_2
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.6.0=py3.7_cuda10.2.89_cudnn7.6.5_0
pytorch-lightning=1.0.2=pypi_0
pytz=2020.1=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=pypi_0
readline=8.0=h7b6447c_0
regex=2020.10.11=pypi_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
s3transfer=0.3.3=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.23.2=py37h0573a6f_0
scikit-multilearn=0.2.0=pypi_0
scipy=1.5.2=py37h0b6359f_0
seaborn=0.11.1=pyhd3eb1b0_0
send2trash=1.5.0=pypi_0
sentencepiece=0.1.91=pypi_0
sentry-sdk=0.19.0=pypi_0
seqeval=1.1.0=pypi_0
setuptools=50.3.0=py37hb0f4dca_1
shortuuid=1.0.1=py37hc8dfbb8_3
simpletransformers=0.60.9=pypi_0
six=1.15.0=py_0
smart_open=2.0.0=py_0
smmap=3.0.4=pyh9f0ad1d_0
sqlite=3.33.0=h62c20be_0
streamlit=0.68.1=pypi_0
subprocess32=3.5.4=pypi_0
tbb=2020.2=hc9558a2_0
tensorboard=2.2.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
tensorboardx=2.1=pypi_0
terminado=0.9.1=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.1=pypi_0
toml=0.10.1=pypi_0
toolz=0.11.1=pypi_0
torchtext=0.7.0=py37
torchvision=0.7.0=py37_cu102
tornado=6.0.4=py37h8f50634_2
tqdm=4.50.2=py_0
traitlets=5.0.4=py_0
transformers=4.3.3=pypi_0
typing-extensions=3.7.4.3=pypi_0
tzlocal=2.1=pypi_0
umap-learn=0.5.0=py37h89c1867_0
urllib3=1.25.10=py_0
validators=0.18.1=pypi_0
wandb=0.10.5=pypi_0
watchdog=0.10.3=py37hc8dfbb8_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yarl=1.6.0=py37h8f50634_0
zipp=3.3.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
